'afterViews'

+++ people/models.py
@@ -3,8 +3,8 @@ from django.db import models
 
 
 class Neighbor(models.Model):
-  x_coord = models.FloatField()
-  y_coord = models.FloatField()
+  x_coord = models.DecimalField(max_digits=5, decimal_places=2)
+  y_coord = models.DecimalField(max_digits=5, decimal_places=2)
   name = models.CharField(max_length=50)
 
   def __str__(self):
